feat: add game modal with game start timer
