+++ apps/client/public/locales/ru/translation.json
@@ -88,6 +88,10 @@
     "red-title": "Команда красных",
     "blue-title": "Команда синих",
     "spymaster": "Спаймастер",
-    "operatives": "Оперативники"
+    "operatives": "Оперативники",
+    "game-start-modal": {
+      "title": "Команды укомплектованы",
+      "text": "Игра начнется через"
+    }
   }
 }
