+++ apps/client/src/app/pages/roomPage/RoomPage.tsx
@@ -1,18 +1,24 @@
 import store from '@/app/store/store';
 import { Pages, TypedSocket } from '@/types/general.types';
 import { RoomState } from '@repo/shared/room';
-import { useEffect } from 'react';
+import { useEffect, useState } from 'react';
 import { useLoaderData, useNavigate } from 'react-router';
 import { RoomTitle } from './roomTitle/RoomTitle';
 import { RoomTeam } from './roomTeam/RoomTeam';
 import { UnknownTeam } from './unknownTeam/UnknownTeam';
+import { ModalWrapper } from '@/app/components/modalWrapper/ModalWrapper';
+import { useTranslation } from 'react-i18next';
+import clsx from 'clsx';
 
 interface ILeftToRoom {
   userId: string;
 }
 
+const onGameStarted = () => console.log('game started');
+
 export function RoomPage() {
   const navigate = useNavigate();
+  const [showModal, setShowModal] = useState(false);
 
   const { roomState, socket } = useLoaderData<{
     roomState: RoomState;
@@ -27,12 +33,36 @@ export function RoomPage() {
       }
     };
 
+    const onStartedGameStartTimer = () => setShowModal(true);
+
     socket.on('lobby:left-room', onLeftRoom);
+    socket.on('room:started-game-start-timer', onStartedGameStartTimer);
+    socket.on('room:started-game', onGameStarted);
 
     return () => {
       socket.off('lobby:left-room', onLeftRoom);
+      socket.off('room:started-game-start-timer', onStartedGameStartTimer);
+      socket.off('room:started-game', onGameStarted);
     };
   });
+
+  const { t } = useTranslation();
+
+  const gameStartModal = (
+    <ModalWrapper>
+      <div
+        className={clsx(
+          'flex flex-col items-center gap-3',
+          'bg-primary-light dark:bg-primary-dark',
+          'p-4 rounded-md border'
+        )}>
+        <h2 className="font-bold text-xl">
+          {t('room.game-start-modal.title')}
+        </h2>
+        <p>{t('room.game-start-modal.text')}</p>
+      </div>
+    </ModalWrapper>
+  );
 
   return (
     <main className="w-full grow flex max-w-7xl flex-col px-3 sm:px-5 gap-2">
@@ -52,6 +82,7 @@ export function RoomPage() {
         />
       </div>
       <UnknownTeam roomState={roomState} socket={socket} />
+      {showModal && gameStartModal}
     </main>
   );
 }
